Move sidebar widgets to generic patient widgets
[sc-27967]
