+++ src/js/views/patients/patient/sidebar/patient-sidebar.scss
@@ -23,45 +23,17 @@
   &:last-of-type {
     padding-bottom: 32px;
   }
-}
 
-.patient-sidebar__divider {
-  border-top: 1px solid $black-90;
-  content: '';
-  display: block;
-  height: 0;
-}
-
-.patient-sidebar__heading {
-  color: $black-60;
-  font-size: 12px;
-  margin-top: 8px;
-
-  &:first-child {
-    margin-top: 0;
+  /* NOTE: context-specific widget styles */
+  .widgets__item + .widgets__item {
+    margin-top: 4px;
   }
-}
 
-.patient-sidebar__item + .patient-sidebar__item {
-  margin-top: 4px;
-}
+  .widgets__heading {
+    margin-top: 8px;
 
-.patient-sidebar__status-inactive {
-  color: $error-color;
-}
-
-.patient-sidebar__status-archived {
-  color: $black-60;
-}
-
-.patient-sidebar__form-widget {
-  width: 100%;
-
-  .fa-expand-alt {
-    font-size: 12px;
-    margin-right: 0;
-    position: absolute;
-    right: 8px;
-    top: 8px;
+    &:first-child {
+      margin-top: 0;
+    }
   }
 }
