+++ src/js/views/patients/patient/sidebar/sidebar_views.js
@@ -1,261 +1,19 @@
-import { bind, each, extend, isFunction, map, propertyOf, reduce } from 'underscore';
+import { bind } from 'underscore';
 import Backbone from 'backbone';
-import { View, CollectionView } from 'marionette';
-import dayjs from 'dayjs';
-import Radio from 'backbone.radio';
-import parsePhoneNumber from 'libphonenumber-js/min';
+import { View } from 'marionette';
 import hbs from 'handlebars-inline-precompile';
 
 import intl from 'js/i18n';
 
 import 'sass/modules/widgets.scss';
 
-import patientTemplate from 'js/utils/patient-template';
-
 import Optionlist from 'js/components/optionlist';
+
+import { WidgetCollectionView } from 'js/views/patients/widgets/widgets_views';
 
 import './patient-sidebar.scss';
 
 const i18n = intl.patients.patient.sidebar.sidebarViews;
-
-// NOTE: widget is a view or view definition
-function buildWidget(widget, patient, widgetModel, options) {
-  if (isFunction(widget)) return new widget(extend({ model: patient }, options, widgetModel.get('definition')));
-
-  return new View(extend({ model: patient }, options, widgetModel.get('definition'), widget));
-}
-
-function getKeyValue(value, keys) {
-  if (keys.length) {
-    return propertyOf(value)(keys);
-  }
-
-  return value;
-}
-
-function getWidgetValue({ fields, name, key, childValue }) {
-  const keys = key && key.split('.') || [];
-
-  if (childValue) {
-    return getKeyValue(childValue, keys);
-  }
-
-  // NOTE: Makes `field_name` optional
-  if (keys.length && !name) name = keys.shift();
-
-  const currentField = fields.find({ name });
-
-  if (!currentField) return;
-
-  return getKeyValue(currentField.get('value'), keys);
-}
-
-// NOTE: These widgets are documented in ./README.md
-const sidebarWidgets = {
-  dob: {
-    modelEvents: {
-      'change:birth_date': 'render',
-    },
-    template: hbs`{{formatHTMLMessage (intlGet "patients.patient.sidebar.sidebarViews.sidebarWidgets.dob") dob=(formatDateTime dob "LONG" inputFormat="YYYY-MM-DD") age=age}}`,
-    templateContext() {
-      const dob = this.model.get('birth_date');
-
-      const age = dayjs().diff(dayjs(dob, 'YYYY-MM-DD'), 'years');
-
-      return { dob, age };
-    },
-  },
-  sex: {
-    modelEvents: {
-      'change:sex': 'render',
-    },
-    template: hbs`{{formatMessage (intlGet "patients.patient.sidebar.sidebarViews.sidebarWidgets.sex") sex=sex}}`,
-  },
-  status: {
-    modelEvents: {
-      'change:status': 'render',
-    },
-    template: hbs`<span class="patient-sidebar__status-{{ status }}">{{formatMessage (intlGet "patients.patient.sidebar.sidebarViews.sidebarWidgets.status") status=status}}</span>`,
-  },
-  divider: {
-    template: hbs`<div class="patient-sidebar__divider"></div>`,
-  },
-  groups: {
-    modelEvents: {
-      'change:_groups': 'render',
-    },
-    template: hbs`{{#each groups}}<div>{{ this.name }}</div>{{/each}}`,
-    templateContext() {
-      return {
-        groups: map(this.model.getGroups().models, 'attributes'),
-      };
-    },
-  },
-  optionsWidget: {
-    className: 'widgets-value',
-    template: hbs`{{ displayValue }}{{#unless displayValue}}{{{ defaultHtml }}}{{/unless}}`,
-    templateContext() {
-      const defaultHtml = this.getOption('default_html');
-      const value = getWidgetValue({
-        fields: this.model.getFields(),
-        name: this.getOption('field_name'),
-        key: this.getOption('key'),
-        childValue: this.getOption('childValue'),
-      });
-      const displayOptions = this.getOption('display_options');
-
-      return {
-        displayValue: displayOptions[value] || value,
-        defaultHtml,
-      };
-    },
-  },
-  phoneWidget: {
-    className: 'widgets-value',
-    template: hbs`{{ displayValue }}{{#unless displayValue}}{{{ defaultHtml }}}{{/unless}}`,
-    templateContext() {
-      const defaultHtml = this.getOption('default_html');
-      const value = getWidgetValue({
-        fields: this.model.getFields(),
-        name: this.getOption('field_name'),
-        key: this.getOption('key'),
-        childValue: this.getOption('childValue'),
-      });
-      if (!value) return { defaultHtml };
-
-      const phone = parsePhoneNumber(value, 'US');
-
-      return {
-        displayValue: phone ? phone.formatNational() : null,
-        defaultHtml,
-      };
-    },
-  },
-  fieldWidget: {
-    className: 'widgets-value',
-    template: hbs`{{ displayValue }}{{#unless displayValue}}{{{ defaultHtml }}}{{/unless}}`,
-    templateContext() {
-      const defaultHtml = this.getOption('default_html');
-      const displayValue = getWidgetValue({
-        fields: this.model.getFields(),
-        name: this.getOption('field_name'),
-        key: this.getOption('key'),
-        childValue: this.getOption('childValue'),
-      });
-      return {
-        displayValue,
-        defaultHtml,
-      };
-    },
-  },
-  templateWidget: View.extend({
-    className: 'widgets-value',
-    initialize() {
-      this.childValue = this.getOption('childValue');
-      this.template = patientTemplate(this.template, this.childValue);
-      this.nestedWidgets = this.template.widgetNames;
-
-      const widgetRegions = reduce(this.nestedWidgets, (regions, widgetName) => {
-        regions[widgetName] = `[data-${ widgetName }-region]`;
-        return regions;
-      }, {});
-
-      this.addRegions(widgetRegions);
-    },
-    serializeData() {
-      return this.model;
-    },
-    onRender() {
-      each(this.nestedWidgets, widgetName => {
-        const widgetModel = Radio.request('entities', 'widgets:model', widgetName);
-        const widget = sidebarWidgets[widgetModel.get('widget_type')];
-
-        this.showChildView(widgetName, buildWidget(widget, this.model, widgetModel, { tagName: 'span', childValue: this.childValue }));
-      });
-    },
-  }),
-  arrayWidget: CollectionView.extend({
-    className: 'widgets-value',
-    childWidget: {
-      widget_type: 'fieldWidget',
-      definition: {},
-    },
-    initialize(options) {
-      const arrayValue = getWidgetValue({
-        fields: this.model.getFields(),
-        name: this.getOption('field_name'),
-        key: this.getOption('key'),
-        childValue: this.getOption('childValue'),
-      });
-
-      each(arrayValue, childValue => {
-        const widgetModel = Radio.request('entities', 'widgets:model', options.child_widget || this.childWidget);
-        const widget = sidebarWidgets[widgetModel.get('widget_type')];
-
-        this.addChildView(buildWidget(widget, this.model, widgetModel, { childValue }));
-      });
-    },
-    template: hbs`{{ defaultHtml }}`,
-    templateContext() {
-      const defaultHtml = this.getOption('default_html');
-
-      return { defaultHtml };
-    },
-  }),
-  formWidget: View.extend({
-    className: 'button-primary patient-sidebar__form-widget',
-    tagName: 'button',
-    template: hbs`{{far "poll-h"}} {{formName}} {{far "expand-alt"}}`,
-    templateContext() {
-      return {
-        formName: this.getOption('form_name'),
-      };
-    },
-    triggers: {
-      'click': 'click',
-    },
-    onClick() {
-      if (this.getOption('is_modal')) {
-        const form = Radio.request('entities', 'forms:model', this.getOption('form_id'));
-        Radio.request('modal', 'show:form', this.model, this.getOption('form_name'), form, this.getOption('modal_size'));
-        return;
-      }
-      Radio.trigger('event-router', 'form:patient', this.model.id, this.getOption('form_id'));
-    },
-  }),
-  dateTimeWidget: {
-    template: hbs`{{formatDateTime dateTime format inputFormat=inputFormat defaultHtml=defaultHtml}}`,
-    templateContext() {
-      const dateTime = getWidgetValue({
-        fields: this.model.getFields(),
-        name: this.getOption('field_name'),
-        key: this.getOption('key'),
-        childValue: this.getOption('childValue'),
-      });
-
-      return {
-        format: this.getOption('format') || 'TIME_OR_DAY',
-        inputFormat: this.getOption('input_format'),
-        dateTime,
-        defaultHtml: this.getOption('default_html'),
-      };
-    },
-  },
-};
-
-const WidgetView = View.extend({
-  className: 'patient-sidebar__section',
-  template: hbs`{{#if definition.display_name}}<div class="patient-sidebar__heading">{{ definition.display_name }}</div>{{/if}}<div class="patient-sidebar__item" data-content-region></div>`,
-  regions: {
-    content: '[data-content-region]',
-  },
-  onRender() {
-    const widget = this.getOption('widget');
-    const patient = this.getOption('patient');
-
-    this.showChildView('content', buildWidget(widget, patient, this.model));
-  },
-});
 
 const NameView = View.extend({
   tagName: 'h1',
@@ -264,23 +22,6 @@ const NameView = View.extend({
   modelEvents: {
     'change': 'render',
   },
-});
-
-const WidgetCollectionView = CollectionView.extend({
-  childView: WidgetView,
-  childViewOptions(model) {
-    const widget = sidebarWidgets[model.get('widget_type')];
-
-    return {
-      widget,
-      model,
-      patient: this.model,
-    };
-  },
-  viewFilter({ model }) {
-    return model.get('widget_type');
-  },
-  viewComparator: false,
 });
 
 const SidebarView = View.extend({
@@ -302,6 +43,7 @@ const SidebarView = View.extend({
     this.showChildView('widgets', new WidgetCollectionView({
       model: this.model,
       collection: this.collection,
+      itemClassName: 'patient-sidebar__section',
     }));
   },
   triggers: {
